pyproject.toml
[tool.poetry]
name = "auth-fastapi"
version = "0.1.0"
description = ""
authors = ["Your Name <34923485+Haste171@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
passlib = "^1.7.4"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
matplotlib = "^3.9.1.post1"
pyjwt = "^2.9.0"
uvicorn = "^0.30.5"
pydantic = {extras = ["email"], version = "^2.8.2"}
bcrypt = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
